refactor(gm): address CodeRabbit review on #521
- Split query.rs (was 512 lines, over the 500 soft cap): moved the mission-query
  handlers (mission_line + handle_mission_list/_full/_details) into the existing
  gm/missions.rs (their natural family); query.rs now 414 lines. Repointed the
  three GM_MISSION_LIST/FULL/DETAILS dispatch arms to missions::.
- show_player_cross_space_and_missing_report_errors: now actually exercises the
  cross-space path (entity in a second space → "different space") in addition to
  the missing-entity path, so the name matches the assertions.
- kill_target / despawn reject tests: added postcondition guards (bystander NPC
  unchanged + no outbound messages) so they fail if a reject path falls through
  to a kill/despawn, instead of only asserting "no panic".

58 gm tests pass; fmt + clippy clean.
